fixed puppet manifest syntax mistakes

+++ 0x0A-configuration_management/2-execute_a_command.pp
@@ -1,4 +1,5 @@
 #kills a process named killmenow
 exec { 'pkill killmenow':
   command => 'pkill killmenow',
+  onlyif  => 'pgrep killmenow'
 }
